split dpgen.py into seperate modules
