+++ .github/workflows/ci.yml
@@ -19,7 +19,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip setuptools wheel
-          pip install flake8 pytest coverage coveralls
+          pip install flake8 pytest pytest-datadir coverage coveralls
           pip install .
       - name: Lint with flake8
         run: |
