+++ miko/tesla/__init__.py
@@ -1,2 +1,2 @@
 from .dpgen import DPTask
-from .screen import SOAPScreening
+# from .screen import SOAPScreening
