+++ tests/tesla/dpgen/data/dpgen_task/record.dpgen
@@ -1,0 +1,5 @@
+0 0
+0 1
+0 2
+0 3
+0 4
